solve TypeError: Cannot handle this data type: (1, 1), <i8 (#7)

+++ pycasia/CASIA.py
@@ -202,6 +202,6 @@ class CASIA:
                 # Comes out as a tuple of chars. Need to be combined. Encoded as gb2312, gotta convert to unicode.
                 label = decode(raw_label[0] + raw_label[1], encoding="gb2312")
                 # Create an array of bytes for the image, match it to the proper dimensions, and turn it into an image.
-                image = fromarray(np.array(photo_bytes).reshape(height, width))
+                image = fromarray(np.array(photo_bytes, dtype=np.uint8).reshape(height, width))
 
                 yield image, label
